+++ project/instance.go
@@ -374,6 +374,15 @@ func instanceNetworkDevices(c *client.Client, p *types.Project, service types.Se
 					}
 					gateway6 = ip.String()
 				}
+
+				_, ok = netConfig.Extensions["ipv4.nat"]
+				if !ok {
+					netConfig.Extensions["ipv4.nat"] = "true"
+				}
+				_, ok = netConfig.Extensions["ipv6.nat"]
+				if !ok {
+					netConfig.Extensions["ipv6.nat"] = "true"
+				}
 			}
 		}
 
@@ -1054,7 +1063,7 @@ func networkExtensions(networkDef types.NetworkConfig) map[string]string {
 	var raw map[string]any
 	ok, err := networkDef.Extensions.Get("x-incus", &raw)
 	if !ok || err != nil || len(raw) == 0 {
-		return nil
+		return map[string]string{}
 	}
 
 	result := make(map[string]string, len(raw))